+++ vortex-app/vortex-app-portal/.env
@@ -1,2 +1,1 @@
-
 CONSOLE_MODULES="examlpe@http://localhost:3003/remoteEntry.js"
