+++ vortex-app/vortex-app-portal/package.json
@@ -52,6 +52,7 @@
     "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^7.1.2",
     "css-minimizer-webpack-plugin": "^7.0.0",
+    "dotenv": "^16.4.5",
     "dotenv-webpack": "^8.1.0",
     "eslint": "^9.12.0",
     "eslint-config-prettier": "^9.1.0",
