+++ vortex-app/vortex-app-portal/src/utils/moduleLoader.js
@@ -2,8 +2,9 @@
 import React from 'react'
 
 function parseModules() {
-  // const modules = process.env.CONSOLE_MODULES
-  const modules = 'vortex_module_template@http://localhost:3003/remoteEntry.js'
+  const modules = process.env.CONSOLE_MODULES
+  // const modules = 'vortex_module_template@http://localhost:3003/remoteEntry.js'
+  console.log('--process.env', process.env.CONSOLE_MODULES)
 
   const parsedModules = []
   if (typeof modules === 'string') {
