+++ vortex-app/vortex-app-portal/src/utils/routers/index.tsx
@@ -2,7 +2,7 @@ import Layout from '@/components/Layout'
 import Dashboard from '@/pages/Dashboard'
 import EdgeModuleContainer from '@/pages/ExampleMicroModule'
 import { createBrowserRouter } from 'react-router-dom'
-import Login from '@/components/AuthProvider/Login'
+import Login from '@/components/Access/Login'
 
 export const router = createBrowserRouter([
   {
